게시판 고치는거 실패.. plan 꾸미기 완료

+++ src/main/java/com/io/threegonew/service/BoardService.java
@@ -21,29 +21,34 @@ public class BoardService {
         return boardRepository.save(request.toEntity());
     }
 
-    // 게시판 글 전체 조회
+    // 게시판 글 전체 조회(최신순 조회)
     public List<Board> findAll() {
         return boardRepository.findAll(Sort.by("bPostdate").descending());
     }
 
-    // 게시판 글 추천순 조회
+//    // 게시판 글 추천순 조회
+//    public List<Board> findAllLikes() {
+//        return boardRepository.findAll(Sort.by("likes").descending());
+//    }
+//
+//    // 게시판 글 조회순 조회
+//    public List<Board> findAllViews() {
+//        return boardRepository.findAll(Sort.by("visitCount").descending());
+//    }
 
-    // 게시판 글 최신순 조회
-
-    // 게시판 글 조회순 조회
-
-    public Board findById(Integer bId) {
+    // 글 아이디로 조회하기
+    public Board findById(Long bId) {
         return boardRepository.findById(bId)
                 .orElseThrow(() -> new IllegalArgumentException("not found : " + bId));
     }
 
     // 게시판 글 삭제
-    public void delete(Integer bId) {
+    public void delete(Long bId) {
         boardRepository.deleteById(bId);
     }
 
     @Transactional
-    public Board update(Integer bId, UpdateBoardRequest request) {
+    public Board update(Long bId, UpdateBoardRequest request) {
         Board board = boardRepository.findById(bId).orElseThrow(() ->
                 new IllegalArgumentException("not found : " + bId));
